+++ pgmax/bp/__init__.py
@@ -1,1 +1,4 @@
 """A sub-package containing functions to perform belief propagation."""
+
+from .bp import BP, decode_map_states, get_marginals
+from .bp_state import BPArrays, BPState, Evidence, FToVMessages, LogPotentials
