+++ pgmax/factor/__init__.py
@@ -5,13 +5,16 @@ from typing import Callable, OrderedDict, Type
 
 import jax.numpy as jnp
 
-from pgmax.factors import enumeration, logical
+from . import enum, logical
+from .enum import EnumerationFactor
+from .factor import Factor, Wiring
+from .logical import LogicalFactor
 
 FAC_TO_VAR_UPDATES: OrderedDict[
     Type, Callable[..., jnp.ndarray]
 ] = collections.OrderedDict(
     [
-        (enumeration.EnumerationFactor, enumeration.pass_enum_fac_to_var_messages),
+        (enum.EnumerationFactor, enum.pass_enum_fac_to_var_messages),
         (logical.ORFactor, logical.pass_logical_fac_to_var_messages),
         (logical.ANDFactor, logical.pass_logical_fac_to_var_messages),
     ]
