--- pgmax/fg/__init__.py
@@ -1,1 +1,0 @@
-"""A sub-package containing classes and functions used to specify Factor Graphs."""
