+++ pgmax/fgraph/__init__.py
@@ -1,0 +1,3 @@
+"""A sub-package containing functions to represent a factor graph."""
+
+from fgraph import FactorGraph, FactorGraphState
