+++ pgmax/fgroup/__init__.py
@@ -1,0 +1,5 @@
+"""A sub-package containing different types of factor groups."""
+
+from .enum import EnumerationFactorGroup, PairwiseFactorGroup
+from .fgroup import FactorGroup, SingleFactorGroup
+from .logical import ANDFactorGroup, ORFactorGroup
