+++ pgmax/fgroup/fgroup.py
@@ -19,105 +19,8 @@ from typing import (
 import jax.numpy as jnp
 import numpy as np
 
-import pgmax.fg.nodes as nodes
+from pgmax import factor
 from pgmax.utils import cached_property
-
-MAX_SIZE = 1e9
-
-
-@total_ordering
-@dataclass(frozen=True, eq=False)
-class VariableGroup:
-    """Class to represent a group of variables.
-    Each variable is represented via a tuple of the form (variable hash, variable num_states)
-
-    Arguments:
-        num_states: An integer or an array specifying the number of states of the variables
-            in this VariableGroup
-    """
-
-    num_states: Union[int, np.ndarray]
-
-    def __post_init__(self):
-        # Only compute the hash once, which is guaranteed to be an int64
-        this_id = id(self) % 2**32
-        _hash = this_id * int(MAX_SIZE)
-        assert _hash < 2**63
-        object.__setattr__(self, "_hash", _hash)
-
-    def __hash__(self):
-        return self._hash
-
-    def __eq__(self, other):
-        return hash(self) == hash(other)
-
-    def __lt__(self, other):
-        return hash(self) < hash(other)
-
-    def __getitem__(self, val: Any) -> Union[Tuple[int, int], List[Tuple[int, int]]]:
-        """Given a variable name, index, or a group of variable indices, retrieve the associated variable(s).
-        Each variable is returned via a tuple of the form (variable hash, variable num_states)
-
-        Args:
-            val: a variable index, slice, or name
-
-        Returns:
-            A single variable or a list of variables
-        """
-        raise NotImplementedError(
-            "Please subclass the VariableGroup class and override this method"
-        )
-
-    @cached_property
-    def variable_hashes(self) -> np.ndarray:
-        """Function that generates a variable hash for each variable
-
-        Returns:
-            Array of variables hashes.
-        """
-        raise NotImplementedError(
-            "Please subclass the VariableGroup class and override this method"
-        )
-
-    @cached_property
-    def variables(self) -> List[Tuple[int, int]]:
-        """Function that returns the list of all variables in the VariableGroup.
-        Each variable is represented by a tuple of the form (variable hash, variable num_states)
-
-        Returns:
-            List of variables in the VariableGroup
-        """
-        assert isinstance(self.variable_hashes, np.ndarray)
-        assert isinstance(self.num_states, np.ndarray)
-        vars_hashes = self.variable_hashes.flatten()
-        vars_num_states = self.num_states.flatten()
-        return list(zip(vars_hashes, vars_num_states))
-
-    def flatten(self, data: Any) -> jnp.ndarray:
-        """Function that turns meaningful structured data into a flat data array for internal use.
-
-        Args:
-            data: Meaningful structured data
-
-        Returns:
-            A flat jnp.array for internal use
-        """
-        raise NotImplementedError(
-            "Please subclass the VariableGroup class and override this method"
-        )
-
-    def unflatten(self, flat_data: Union[np.ndarray, jnp.ndarray]) -> Any:
-        """Function that recovers meaningful structured data from internal flat data array
-
-        Args:
-            flat_data: Internal flat data array.
-
-        Returns:
-            Meaningful structured data
-        """
-        raise NotImplementedError(
-            "Please subclass the VariableGroup class and override this method"
-        )
 
 
 @total_ordering
@@ -208,7 +111,7 @@ class FactorGroup:
         return factor_edges_num_states
 
     @cached_property
-    def _variables_to_factors(self) -> Mapping[FrozenSet, nodes.Factor]:
+    def _variables_to_factors(self) -> Mapping[FrozenSet, factor.Factor]:
         """Function to compile potential array for the factor group.
         This function is only called on demand when the user requires it.
 
@@ -223,7 +126,7 @@ class FactorGroup:
         return self.log_potentials.flatten()
 
     @cached_property
-    def factors(self) -> Tuple[nodes.Factor, ...]:
+    def factors(self) -> Tuple[factor.Factor, ...]:
         """Returns all factors in the factor group.
         This function is only called on demand when the user requires it."""
         return tuple(self._variables_to_factors.values())
@@ -304,7 +207,7 @@ class SingleFactorGroup(FactorGroup):
         factor: the single factor in the SingleFactorGroup
     """
 
-    factor: nodes.Factor
+    factor: factor.Factor
 
     def __post_init__(self):
         super().__post_init__()
@@ -329,7 +232,7 @@ class SingleFactorGroup(FactorGroup):
 
     def _get_variables_to_factors(
         self,
-    ) -> OrderedDict[FrozenSet, nodes.Factor]:
+    ) -> OrderedDict[FrozenSet, factor.Factor]:
         """Function that generates a dictionary mapping names to factors.
 
         Returns:
