+++ pgmax/fgroup/logical.py
