+++ pgmax/vgroup/__init__.py
@@ -1,0 +1,5 @@
+"""A sub-package containing different types of variable groups."""
+
+from .varray import NDVariableArray
+from .vdict import VariableDict
+from .vgroup import VariableGroup
